requirements.txt
segmentation-models-pytorch
imgaug
pytorch_lightning